Added extra validation and started the testing projects

+++ Brighthouse.News.Api/Features/ArticleManage/ArticleManageService.cs
@@ -4,20 +4,21 @@ using Brighthouse.News.Api.Features.ArticleDisplay;
 using Brighthouse.News.Api.Infrastructure.Repositories;
 using Brighthouse.News.Api.Models;
 using FluentValidation;
+using FluentValidation.Results;
 using Mapster;
 
 namespace Brighthouse.News.Api.Features.ArticleManage
 {
     public class ArticleManageService(ILogger<ArticleDisplayService> logger,
-        NewsRepository newsRepository, IValidator<ArticleAddDto> addValidator,
-        IValidator<ArticleUpdateDto> updateValidator, IValidator<ArticleDeleteDto> deleteValidator) : IArticleManageService
+        INewsRepository newsRepository, IValidator<ArticleAddDto> addValidator,
+        IValidator<ArticleUpdateDto> updateValidator, IValidator<ArticleDetailInputDto> deleteValidator) : IArticleManageService
     {
 
         private readonly ILogger<ArticleDisplayService> _logger = logger;
-        private readonly NewsRepository _newsRepository = newsRepository;
+        private readonly INewsRepository _newsRepository = newsRepository;
         private readonly IValidator<ArticleAddDto> _addValidator = addValidator;
         private readonly IValidator<ArticleUpdateDto> _updateValidator = updateValidator;
-        private readonly IValidator<ArticleDeleteDto> _deleteValidator = deleteValidator;
+        private readonly IValidator<ArticleDetailInputDto> _deleteValidator = deleteValidator;
 
         public async Task<BrighthouseResponse> AddArticleAsync(ArticleAddDto input)
         {
@@ -32,15 +33,28 @@ namespace Brighthouse.News.Api.Features.ArticleManage
                     var article = new Article();
                     article = input.Adapt<Article>();
 
-                    var articleId = await _newsRepository.AddArticleAsync(article);
+                    var duplicateArticle = await _newsRepository.GetArticleByTitleAndAuthorAsync(input.AuthorId, input.Title);
 
-                    response.Success = true;
-                    response.Message = "The article has been successfully added";
-                    response.Data = articleId;
+                    if(duplicateArticle != null)
+                    {
+                        validationResult.Errors.Add(new ValidationFailure(nameof(input.Title), "An article already exists for this author with this title"));
+
+                        response.Success = false;
+                        response.Message = "The article could not be added because of validation errors";
+                        response.ValidationResult = validationResult;
+                    }
+                    else
+                    {
+                        var articleId = await _newsRepository.AddArticleAsync(article);
+
+                        response.Success = true;
+                        response.Message = "The article has been successfully added";
+                        response.Data = articleId;
+                    }
                 }
                 else
                 {
-                    response.Success = false; ;
+                    response.Success = false;
                     response.Message = "The article could not be added because of validation errors";
                     response.ValidationResult = validationResult;
                 }
@@ -68,14 +82,27 @@ namespace Brighthouse.News.Api.Features.ArticleManage
                     var article = new Article();
                     article = input.Adapt<Article>();
 
-                    await _newsRepository.UpdateArticleAsync(article);
+                    var existingArticle = await _newsRepository.GetArticleAsync(input.Id);
 
-                    response.Success = true;
-                    response.Message = "The article has been successfully updated";
+                    if (article == null)
+                    {
+                        validationResult.Errors.Add(new ValidationFailure(nameof(input.Id), "Article does not exist"));
+
+                        response.Success = false;
+                        response.Message = "The article could not be updated because of validation errors";
+                        response.ValidationResult = validationResult;
+                    }
+                    else
+                    {
+                        await _newsRepository.DeleteArticleAsync(input.Id);
+
+                        response.Success = true;
+                        response.Message = "The article has been successfully updated";
+                    }
                 }
                 else
                 {
-                    response.Success = false; ;
+                    response.Success = false;
                     response.Message = "The article could not be updated because of validation errors";
                     response.ValidationResult = validationResult;
                 }
@@ -90,7 +117,7 @@ namespace Brighthouse.News.Api.Features.ArticleManage
             return response;
         }
 
-        public async Task<BrighthouseResponse> DeleteArticleAsync(ArticleDeleteDto input)
+        public async Task<BrighthouseResponse> DeleteArticleAsync(ArticleDetailInputDto input)
         {
             var response = new BrighthouseResponse();
 
@@ -100,14 +127,27 @@ namespace Brighthouse.News.Api.Features.ArticleManage
 
                 if (validationResult.IsValid)
                 {
-                    await _newsRepository.DeleteArticleAsync(input.ArticleId);
+                    var article = await _newsRepository.GetArticleAsync(input.ArticleId);
 
-                    response.Success = true;
-                    response.Message = "The article has been successfully deleted";
+                    if(article == null)
+                    {
+                        validationResult.Errors.Add(new ValidationFailure(nameof(input.ArticleId), "Article does not exist"));
+
+                        response.Success = false;
+                        response.Message = "The article could not be deleted because of validation errors";
+                        response.ValidationResult = validationResult;
+                    }
+                    else
+                    {
+                        await _newsRepository.DeleteArticleAsync(input.ArticleId);
+
+                        response.Success = true;
+                        response.Message = "The article has been successfully deleted";
+                    }
                 }
                 else
                 {
-                    response.Success = false; ;
+                    response.Success = false;
                     response.Message = "The article could not be deleted because of validation errors";
                     response.ValidationResult = validationResult;
                 }
